apps/DjangoApi/requirements.txt
django==5.0.1
djangorestframework==3.14.0
psycopg[binary,pool]==3.2.12
gunicorn==21.2.0
python-dotenv==1.0.0
orjson==3.9.15
pymongo>=4.10
